requirements.txt
langchain
python-dotenv
openai
rich